+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and (probably only like BooleanState cluster)
 - [matterbridgeDevice]: Added waterFreezeDetector, waterLeakDetector, rainSensor, smokeCoAlarm, electricalSensor and deviceEnergyManagement device types as conformance to Matter 1.3
 - [matterbridgeDevice]: Added all clusters needed for the above Matter 1.3 device types 
 - [matterbridgeDevice]: Added FanControl cluster (rev. 2) helper methods for the Fan device type
+- [matterbridge]: Added parameter -matterlogger [debug | info | notice | warn | error | fatal] to set the matter.js Logger separately from the Matterbridge log
 
 <a href="https://www.buymeacoffee.com/luligugithub">
   <img src="./yellow-button.png" alt="Buy me a coffee" width="120">